# backend/internal/apidocs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Yivi Business Wallet API
  description: |
    HTTP API for the Yivi Business Wallet backend. All endpoints live under the
    versioned `/api/v1` prefix. Authentication uses the `ybw_session` cookie set
    by the login flow; endpoints that do not require it are marked with no security.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Versioned API root
...
paths:
...
  /api/v1/proofing/flows:
    get:
      tags: [identity-proofing]
      summary: List the customer's flows (public API)
      description: >-
        Authenticated with a customer API key as a Bearer token, not a session. The
        flows assigned to the key's customer that a subject can finish, with the default
        one marked.
      security:
        - customerApiKey: []
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/proof/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    get:
      tags: [identity-proofing]
      summary: Open a hosted proofing link
      security: []
      description: >-
        The hosted page, unauthenticated: the token is the whole authorization, and every route is rate-limited per link. The customer's `name` and `branding` (its `logoUri` is this link's logo route), what the flow collects (`requestedAttributes`, `requiredAssuranceLevel`, `yiviAvailable`), and the progress: `status`, `started`, and `linkExpiresAt` (the link's end until started, then the session's). `sessionId` is the customer's id for the session; `redirectUrl` (absent for the page's own done screen) where the page sends the subject once settled; `embedOrigins` the customer's allowed redirect origins, the only ones the page posts its `proofing.completed` message to; `language` (`en` or `nl`, absent for the browser's) the page's language, and `locales` the languages the flow's page offers (empty: every one). No personal data.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/v1/proofing/sessions:
    get:
      tags: [identity-proofing]
      summary: List proofing sessions (public API)
      description: >-
        Authenticated with a customer API key with `sessions:read`. The customer's
        sessions, newest first, as stored (no re-check at the identity-proofing-service),
        `limit` per page (default 50, at most 100): `{sessions, nextCursor}`. Pass
        `nextCursor` back as `cursor` for the next page; it is null on the last.
      security:
        - customerApiKey: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    post:
      tags: [identity-proofing]
      summary: Create a proofing session (public API)
      description: >-
        Send an `Idempotency-Key` header to make a retry safe: the same key and body
        within 24 hours get the first answer again (`Idempotent-Replayed: true`), the
        key with another body is 422 `idempotency_key_reused`, and while the first call
        still runs 409 `idempotency_in_flight`. `cancel` takes the header too.
        Authenticated with a customer API key as a Bearer token. `email` and an optional
        `name` of the subject, `flowId` (absent is the customer's default flow) and
        `sendMail` (absent or true mails the subject as a member's request would), and
        `language` (`en` or `nl`: the mail's and the Idem app's; absent is the deployment
        default for the mail and the phone's language in the app). The answer carries the session's vcmrtd `deepLink`, for a caller that shows the QR code
        itself, and `mailSent`. Refused while the customer is paused. With a test key
        the session runs in the org's sandbox and resolves at once to `scriptedOutcome`
        (`approve`, the default; `reject:<CODE>`; `needs_review`; or `expire`): it is not
        mailed, has no `deepLink`, is left out of the stats, and carries `livemode:
        false`, as do its webhooks. A live key may not set `scriptedOutcome` (400).
        With `hosted: true` no session is created yet: the answer carries `hostedUrl`,
        a link to the public page the subject opens on their own device (valid 72 hours)
        and starts once from; nothing is mailed. Not with a test key. `redirectUrl`, only with
        `hosted: true`, is where the page sends the subject once the session settles, with
        `?session=<id>&status=<status>` added; its origin must be one of the customer's
        `allowedRedirectOrigins` (else 400 `redirect_not_allowed`).
      security:
        - customerApiKey: []
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: Proofing is paused for the customer, or the deployment has no encryption key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: The flow does not exist, cannot be completed in the app, or is not assigned to the customer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/proofing/sessions/{sessionID}:
    parameters:
      - name: sessionID
        in: path
        required: true
        schema:
          type: string
          pattern: '^ps_[a-z2-7]{26}$'
        description: Session id (`ps_…`), as the create answer and webhooks give it.
    get:
      tags: [identity-proofing]
      summary: Get a proofing session (public API)
      description: >-
        Authenticated with a customer API key as a Bearer token. One of the customer's
        sessions, re-checked at the identity-proofing-service first while it may still
        change: its `status`, assurance, `errorCode`, and for an approved subject the name
        read off the document (`proofedName`) until the customer's data retention clears
        it.
      security:
        - customerApiKey: []
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [identity-proofing]
      summary: Erase a proofing session (public API)
      description: >-
        Authenticated with a customer API key as a Bearer token. Erases the session at
        the identity-proofing-service and what the wallet holds of its outcome (the
        proofed name), whatever its state, and sends `session.purged`. The session stays
        readable with `purgedAt` set, for the audit trail. Audited
        `identity_proofing.session_purged`; erasing it again changes nothing.
      security:
        - customerApiKey: []
      responses:
        '204':
          description: Erased.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/proofing/sessions/{sessionID}/result:
    parameters:
      - name: sessionID
        in: path
        required: true
        schema:
          type: string
          pattern: '^ps_[a-z2-7]{26}$'
        description: Session id (`ps_…`), as the create answer and webhooks give it.
    get:
      tags: [identity-proofing]
      summary: Get a proofing session's verified identity (public API)
      description: >-
        Authenticated with a customer API key with the `results:read` scope. A settled
        session's result, read from the identity-proofing-service on each call (the
        wallet stores no identity): `status`, `method`, assurance, `verifiedAt`, for an
        approved session `identity` (`givenName`, `familyName`, `birthDate`,
        `nationality`), and `evidence` (`type` `emrtd` or `yivi_disclosure`, the
        document's type, issuing state and expiry, `passiveAuth` and `activeAuth`
        `valid`/`invalid`/`not_performed`, `faceMatch`, `liveness`). Never the document
        number or an image. Every read is audited `identity_proofing.result_read`.
      security:
        - customerApiKey: []
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The session has no outcome yet (`result_not_ready`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/proofing/sessions/{sessionID}/methods/{method}:
    parameters:
      - name: sessionID
        in: path
        required: true
        schema:
          type: string
          pattern: '^ps_[a-z2-7]{26}$'
        description: Session id (`ps_…`), as the create answer and webhooks give it.
      - name: method
        in: path
        required: true
        schema:
          type: string
          enum: [idem_app, yivi_app]
        description: The app the subject proofs with.
    post:
      tags: [identity-proofing]
      summary: Start a hosted session in an app, headless (public API)
      description: >-
        Authenticated with a customer API key with `sessions:write`, for a customer
        building its own UI: starts a session created with `hosted: true` in the app, as
        the subject picking it on the hosted page would, and answers the link to show:
        `appLink` (the Idem app's `vcmrtd://` link) or `walletLink` (the Yivi app's
        `openid4vp://` request, its disclosure started already), with `expiresAt`. The
        Yivi face check then runs through `/api/v1/proof/{token}/yivi/…` with the token
        of the `hostedUrl`, which also serves `claim-link` for a new Idem code. Takes an
        `Idempotency-Key`.
      security:
        - customerApiKey: []
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The session was not created hosted (`not_hosted`), was started already, or ended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    customerApiKey:
      type: http
      scheme: bearer
      description: A proofing customer's API key (`yp_live_…`), for the public `/api/v1/proofing` routes.
    cookieAuth:
      type: apiKey
      in: cookie
      name: ybw_session
  schemas:
    Error:
      type: object
      description: Error response body returned by respond.Error.
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Stable machine-readable error code.
      required: [error, code]
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No valid session cookie was supplied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The session lacks the required role for this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The referenced resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: >-
        The customer spent its public API budget (code `rate_limited`): 120 calls a
        minute, and of those 10 session creations. `Retry-After` says in how many
        seconds to try again.
      headers:
        Retry-After:
          description: Seconds until the next call is allowed.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - cookieAuth: []